personalized-bot/requirements.txt
bubbletea-chat[llm]
python-dotenv
